setting up for github workflow
