+++ demoplswork/src/test/java/MaterialTest.java
@@ -28,7 +28,7 @@ public class MaterialTest {
     public void testPositive1() { assertTrue(material.getQuantity() > 0); }
 
     @Test
-    public void getCost() { assertEquals(3.5, material.getPrice()); }
+    public void getPrice() { assertEquals(3.5, material.getPrice()); }
     @Test
     public void testPositive2() { assertTrue(material.getPrice() > 0); }
